pyproject.toml
[tool.poetry]
name = "project-api-tests"
version = "0.1.0"
description = ""
authors = ["viktoriya <tkvika@yandex.ru>"]
readme = "README.md"
packages = [{include = "project_api_tests"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
pytest = "^7.1.3"
voluptuous = "^0.13.1"
pytest-voluptuous = "^1.2.0"
selene = {version = "^2.0.0-beta.8", allow-prereleases = true}
allure-pytest = "^2.10.0"
packaging = "^21.3"
curlify = "^2.2.1"
python-dotenv = "^0.21.0"

[tool.pytest.ini_options]
addopts = "--alluredir allure-results --clean-alluredir "

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
